% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{agree_to_upcoming_loggit_updates}
\alias{agree_to_upcoming_loggit_updates}
\title{Agree to upcoming loggit changes}
\usage{
agree_to_upcoming_loggit_updates()
}
\description{
This function takes no arguments, but must be run before calling \code{loggit()}
in a session. Due to the large, breaking changes coming in loggit v2.0, users
are required to acknowledge that they have seen the error message before
continuing to use this package's logging facilities.
}
