% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{default_ndjson_sanitizer}
\alias{default_ndjson_sanitizer}
\title{Default sanitizer for ndJSON.}
\usage{
default_ndjson_sanitizer(string)
}
\description{
This is the default ndJSON sanitizer function for incoming log data. This
type of function is needed because since \code{loggit} reimplements its own
string-based JSON parser, and not a fancy one built from an AST or something,
it's very easy to have bad patterns break your logs.
}
