% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{get_rotate_lines}
\alias{get_rotate_lines}
\title{Get Log File Rotation}
\usage{
get_rotate_lines()
}
\description{
Get the configuration for the number of lines after which log files are
truncated.
}
\examples{
get_rotate_lines()

}
