% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{get_timestamp_format}
\alias{get_timestamp_format}
\title{Get Timestamp Format}
\usage{
get_timestamp_format()
}
\description{
Get timestamp format for use in output logs.
}
\examples{
get_timestamp_format()

}
