% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{getTimestampFormat}
\alias{getTimestampFormat}
\title{Get Timestamp Format}
\usage{
getTimestampFormat()
}
\description{
Get timestamp format for use in output logs.
}
\examples{
getTimestampFormat()

}
