% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggit.R
\name{loggit}
\alias{loggit}
\title{Log Message to Output File}
\usage{
loggit(log_lvl, log_msg, log_detail = "")
}
\arguments{
\item{log_lvl}{Level of log output. In actual practice, one of "INFO",
"WARN", and "STOP" ("ERROR") are common.}

\item{log_msg}{Main log message.}

\item{log_detail}{Additional detail recored along with a log message.}
}
\description{
This function executes immediately before the function definitions for the
base handler functions (\code{\link[base]{message}},
\code{\link[base]{warning}}, and
\code{\link[base]{stop}}/\code{\link[base]{stopifnot}}), and logs their
timestamped output (a bit more verbosely) to a log file. The log file
defaults to a \href{https://www.json.org/}{JSON} file, which is a portable
file format that is easily parsed by many systems, but will eventually have a
\code{.txt} option as well.
}
\details{
While this function has an intended use of logging handler messages without
any direct user interaction, it is flexible enough to be used as you see fit.
}
