% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggit.R
\name{loggitall}
\alias{loggitall}
\title{Log It All!}
\usage{
loggitall(pkgs)
}
\arguments{
\item{pkgs}{Character vector of package names, or something that can be
coerced to a character vector.}
}
\description{
This function creates backups of loaded R package \code{NAMESPACE} files, and
adds the `loggit` package as an additional import to the package. On R
termination, the backups are restored. This allows `loggit` to mask the base
R condition handlers in all package functions, and not just those that you
define or call.
}
\details{
Note: I don't know if the original \code{NAMESPACE} files will be restored if
R crashes before this finishes executing. Use at your own risk of
reinstalling your packages!
}
