% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{message}
\alias{message}
\title{Diagnostic Messages, and Output to Log}
\usage{
message(..., domain = NULL, appendLF = TRUE, .loggit = TRUE,
  log_detail = "", echo = FALSE)
}
\arguments{
\item{...}{Same as base.}

\item{domain}{Same as base.}

\item{appendLF}{Same as base.}

\item{.loggit}{Should loggit function execute? Defaults to \code{TRUE}.}

\item{log_detail}{Additional information to accompany log message. Must be a
string.}

\item{echo}{Should loggit's log be echoed to the console, as well? Defaults
to \code{FALSE}.}
}
\description{
Identical to base R's \code{\link[base]{message}}, but with included message
logging.
}
