% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{read_ndjson}
\alias{read_ndjson}
\title{Read ndJSON-formatted log file}
\usage{
read_ndjson(logfile)
}
\arguments{
\item{logfile}{Log file to read from, and convert to a \code{data.frame}.}
}
\value{
A \code{data.frame}
}
\description{
Read ndJSON-formatted log file
}
