% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{set_rotate_lines}
\alias{set_rotate_lines}
\title{Set Log File Rotation}
\usage{
set_rotate_lines(rotate_lines = NULL, confirm = TRUE)
}
\arguments{
\item{rotate_lines}{The number of log entries to keep in the logfile.
Defaults to \code{NULL}, see \code{rotate_logs()} for implementation.}

\item{confirm}{Print confirmation message of timestamp format? Defaults to
\code{TRUE}.}
}
\description{
Set configuration for the number of lines after which log files are.
truncated. The function will echo out the currently set line number if
is set, or else it will echo that there is no limit set.
}
\details{
This function provides a means of setting automatic log file truncation.
Log files are not truncated when the value is set to \code{NULL}, which is
also the default setting.
}
\examples{
set_rotate_lines(100)  # set limit to 100 lines
set_rotate_lines(NULL) # turn off auto-truncation

}
