% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{set_timestamp_format}
\alias{set_timestamp_format}
\title{Set Timestamp Format}
\usage{
set_timestamp_format(ts_format = "\%Y-\%m-\%dT\%H:\%M:\%S\%z", confirm = TRUE)
}
\arguments{
\item{ts_format}{ISO date format. Defaults to ISO-8601 (e.g.
"2020-01-01T00:00:00+0000").}

\item{confirm}{Print confirmation message of timestamp format? Defaults to
\code{TRUE}.}
}
\description{
Set timestamp format for use in output logs. This function performs no time
format validations, but will echo out the current time in the provided format
for manual validation.
}
\details{
This function provides no means of setting a timezone, and instead relies on
the host system's time configuration to provide this. This is to enforce
consistency across software running on the host.
}
\examples{
set_timestamp_format("\%Y-\%m-\%d \%H:\%M:\%S")

}
