% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{setLogFile}
\alias{setLogFile}
\title{Set Log File}
\usage{
setLogFile(logfile = "loggit.json", folder = getwd())
}
\arguments{
\item{logfile}{Name of log file.}

\item{folder}{Folder in which the log file should exist.}
}
\description{
Set the log file that loggit will write to. Defaults to one called
"loggit.json" in the working directory that exists on load.
}
