% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{stopifnot}
\alias{stopifnot}
\title{Ensure the Truth of R Expressions, and Output to Log}
\usage{
stopifnot(..., log_detail = "")
}
\arguments{
\item{...}{Same as base.}

\item{log_detail}{Additional information to accompany log message. Must be a
string.}
}
\description{
Identical to base R's \code{\link[base]{stopifnot}}, but with included
message logging. Also includes an added argument for control over additional
log detail in log output.
}
