% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{warning}
\alias{warning}
\title{Warning Messages, and Output to Log}
\usage{
warning(..., call. = TRUE, immediate. = FALSE, noBreaks. = FALSE,
  domain = NULL, .loggit = TRUE)
}
\arguments{
\item{...}{Same as base.}

\item{call.}{Same as base.}

\item{immediate.}{Same as base.}

\item{noBreaks.}{Same as base.}

\item{domain}{Same as base.}

\item{.loggit}{Should loggit function execute? Defaults to \code{TRUE}.}
}
\description{
Identical to base R's \code{\link[base]{warning}}, but with included message
logging.
}
