% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{write_ndjson}
\alias{write_ndjson}
\title{Write ndJSON-formatted log file}
\usage{
write_ndjson(log_df, echo = TRUE, overwrite = FALSE)
}
\arguments{
\item{overwrite}{}
}
\value{

}
\description{
Write ndJSON-formatted log file
}
