% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{message}
\alias{message}
\title{Diagnostic Messages, and Output to Log}
\usage{
message(..., domain = NULL, appendLF = TRUE)
}
\arguments{
\item{...}{Same as base.}

\item{domain}{Same as base.}

\item{appendLF}{Same as base.}
}
\description{
Identical to base R's \code{\link[base]{message}}, but with included message
logging.
}
