% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{setLogFile}
\alias{setLogFile}
\title{Set Log File}
\usage{
setLogFile(logfile = paste0(getwd(), "/", "loggit.json"), confirm = TRUE)
}
\arguments{
\item{logfile}{Full path to log file. Until other output formats are
introduced, the file name must end in ".json".}

\item{confirm}{Print confirmation of log file setting? Defaults to \code{TRUE}.}
}
\description{
Set the log file that loggit will write to. Defaults to one called
"loggit.json" in the working directory that exists on load.
}
