% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{setTimeStampFormat}
\alias{setTimeStampFormat}
\title{Set Timestamp Format}
\usage{
setTimeStampFormat(ts_format = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{ts_format}{ISO date format.}
}
\description{
Set timestamp format for use in output logs.
}
