% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{setTimestampFormat}
\alias{setTimestampFormat}
\title{Set Timestamp Format}
\usage{
setTimestampFormat(ts_format = "\%Y-\%m-\%d \%H:\%M:\%S", confirm = TRUE)
}
\arguments{
\item{ts_format}{ISO date format.}

\item{confirm}{Print confirmation of timestamp format setting? Defaults to
\code{TRUE}.}
}
\description{
Set timestamp format for use in output logs.
}
\examples{
setTimestampFormat("\%Y-\%m-\%d \%H:\%M:\%S")

}
