% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{warning}
\alias{warning}
\title{Warning Messages, and Output to Log}
\usage{
warning(..., call. = TRUE, immediate. = FALSE, noBreaks. = FALSE,
  domain = NULL, .loggit = TRUE, log_detail = "", echo = FALSE)
}
\arguments{
\item{...}{Same as base.}

\item{call.}{Same as base.}

\item{immediate.}{Same as base.}

\item{noBreaks.}{Same as base.}

\item{domain}{Same as base.}

\item{.loggit}{Should loggit function execute? Defaults to \code{TRUE}.}

\item{log_detail}{Additional information to accompany log message. Must be a
string.}

\item{echo}{Should loggit's log be echoed to the console, as well? Defaults
to \code{FALSE}.}
}
\description{
Identical to base R's \code{\link[base]{warning}}, but with included message
logging.
}
